requirements.txt
pdfplumber
wordcloud
python-docx
